modificado código para correr y funciona bien

+++ signjoey/search.py
@@ -359,7 +359,7 @@ def beam_search(
         batch_index = topk_beam_index + beam_offset[
             : topk_beam_index.size(0)
         ].unsqueeze(1)
-        select_indices = batch_index.view(-1)
+        select_indices = batch_index.view(-1).long()
 
         # append latest prediction
         alive_seq = torch.cat(
@@ -414,7 +414,7 @@ def beam_search(
             )
 
         # reorder indices, outputs and masks
-        select_indices = batch_index.view(-1)
+        select_indices = batch_index.view(-1).long()
         encoder_output = encoder_output.index_select(0, select_indices)
         src_mask = src_mask.index_select(0, select_indices)
